refactor: move ProjectSrttingsScreen
